+++ src/main/java/holiday/InMemoryHoursTaken.java
@@ -7,22 +7,22 @@ import javax.inject.Singleton;
 
 @Singleton
 public class InMemoryHoursTaken implements HoursTaken {
-	private final Map<String,Integer> hoursOff = new ConcurrentHashMap<>();
-	private final Map<String,Integer> vacationHours = new ConcurrentHashMap<>();
+	private final Map<String,Double> hoursOff = new ConcurrentHashMap<>();
+	private final Map<String,Double> vacationHours = new ConcurrentHashMap<>();
 	@Override
-	public void takeHoursOff(int hours, String month) {
+	public void takeHoursOff(double hours, String month) {
 		hoursOff.put(month, hours);
 	}
 	@Override
-	public void takeVacationHours(int hours, String month) {
+	public void takeVacationHours(double hours, String month) {
 		vacationHours.put(month, hours);
 	}
 	@Override
-	public int hoursOffTaken(String month) {
+	public double hoursOffTaken(String month) {
 		return hoursOff.containsKey(month) ? hoursOff.get(month) : 0;
 	}
 	@Override
-	public int vacationHoursTaken(String month) {
+	public double vacationHoursTaken(String month) {
 		return vacationHours.containsKey(month) ? vacationHours.get(month) : 0;
 	}
 }
